requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
appdirs==1.4.4
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
backcall==0.2.0
blinker==1.6.2
bootstrap4==0.1.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
databricks-cli==0.17.7
debugpy==1.6.5
decorator==5.1.1
Django==4.2
django-cors-headers==4.0.0
django-webpack-loader==2.0.0
docker==6.1.2
entrypoints==0.4
executing==1.2.0
ffmpeg==1.4
ffmpeg-downloader==0.2.0
Flask==2.3.2
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-speech==2.20.0
google-cloud-storage==2.9.0
google-cloud-texttospeech==2.14.1
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.49.1
grpcio-status==1.54.2
grpcio-tools==1.49.1
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.19.4
ipython==8.8.0
itsdangerous==2.1.2
jax==0.4.10
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.2
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.40.1rc1
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mlflow==2.3.0
msgpack==1.0.5
multidict==6.0.4
mysqlclient==2.1.1
nest-asyncio==1.5.6
numba==0.57.0
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.6
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==22.0
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
platformdirs==2.6.2
pooch==1.6.0
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==4.21.6
psutil==5.9.4
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.11
pycparser==2.21
pydub==0.25.1
Pygments==2.14.0
PyJWT==2.7.0
PyMySQL==1.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
querystring-parser==1.2.4
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
setuptools==66.0.0
six==1.16.0
smmap==5.0.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.14
sqlparse==0.4.4
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.65.0
traitlets==5.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
waitress==2.1.2
wcwidth==0.2.5
websocket-client==1.5.1
Werkzeug==2.3.4
wheel==0.38.4
wrapt==1.14.1
yarl==1.9.2
zipp==3.15.0
